requirements.txt
PySide6==6.9.2; python_version >= "3.13"
PySide6==6.7.2; python_version < "3.13"
psutil==6.0.0
WMI==1.5.1
pynvml==11.5.0
mss==9.0.1
dxcam==0.0.5
numpy>=2.1.3
speedtest-cli==2.1.3
PyYAML==6.0.2
pywin32>=306
keyboard>=0.13.5